requirements.txt
protobuf==3.20.*
numpy==1.19.5
h5py==2.10.0
kafka-python==2.0.2
tensorflow==2.4.0
pandas==1.2.3
scipy==1.6.1
scikit_learn==1.0.1
matplotlib==3.3.4
influxdb_client==1.20.0